+++ src/interpreter/rank_order.h
@@ -25,7 +25,7 @@ class rank_order_int : public interpreter {
 	public:
 		bool is_this_format(const vector<string> & inputs) const;
 
-		pair<map<int, string>, list<ballot_group> > interpret_ballots(
+		pair<map<size_t, string>, list<ballot_group> > interpret_ballots(
 		    const vector<string> & inputs,
 		    bool debug) const;
